requirements.txt
python-dotenv
vmwc
ldap3
pywinrm
pymongo==3.12.1
pythonping
pytz
pyOpenSSL
ndg-httpsclient
pyasn1
pymongo[srv]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas
boto3
openai
streamlit
matplotlib
wordcloud
great-expectations
python-dotenv
requests
